Topology-lift models, transactions, claims, tasks, runs, events, artifacts, finalization, and notifications (#82591)
